backend/requirements.txt
flask
easyocr
numpy
pdf2image
PyMuPDF
supabase
python-dotenv
openai